Fixed copy and paste bugs
Added "pos" arg to paste so we can control the position. There seems to
be some confusion with par and new_par in the core plugin. I've changed
the callback function to send the parent instead of the new_parent to
the delta updates so the copy goes through correctly.